Azure VM cleanup added in deployment workflow
Azure VM cleanup added in CD deployment workflow.
Now the workflow starts with cleaning up the azure vm then copy the required files and installs the helm charts. Validates the CD job be checking the status of the helm deployment.
If the helm charts is in deployed state it is considered as a success

Signed-off-by: jothikaaravindhan@gmail.com <jothikaaravindhan@gmail.com>